+++ source/MaterialXGenShader/Util.cpp
@@ -699,101 +699,4 @@ ValueElementPtr findNodeDefChild(const string& path, DocumentPtr doc, const stri
     return valueElement;
 }
 
-unsigned int getUIProperties(const ValueElementPtr nodeDefElement, UIProperties& uiProperties)
-{
-    if (!nodeDefElement)
-    {
-        return 0;
-    }
-
-    unsigned int propertyCount = 0;
-    uiProperties.uiName = nodeDefElement->getAttribute(ValueElement::UI_NAME_ATTRIBUTE);
-    if (!uiProperties.uiName.empty())
-        propertyCount++;
-
-    uiProperties.uiFolder = nodeDefElement->getAttribute(ValueElement::UI_FOLDER_ATTRIBUTE);
-    if (!uiProperties.uiFolder.empty())
-        propertyCount++;
-
-    if (nodeDefElement->isA<Parameter>())
-    {
-        string enumString = nodeDefElement->getAttribute(ValueElement::ENUM_ATTRIBUTE);
-        if (!enumString.empty())
-        {
-            uiProperties.enumeration = splitString(enumString, ",");
-            if (uiProperties.enumeration.size())
-                propertyCount++;
-        }
-
-        const string& enumerationValues = nodeDefElement->getAttribute(ValueElement::ENUM_VALUES_ATTRIBUTE);
-        if (!enumerationValues.empty())
-        {
-            const string& elemType = nodeDefElement->getType();
-            const TypeDesc* typeDesc = TypeDesc::get(elemType);
-            if (typeDesc->isScalar() || typeDesc->isFloat2() || typeDesc->isFloat3() || 
-                typeDesc->isFloat4())
-            {
-                StringVec stringValues = splitString(enumerationValues, ",");
-                string valueString;
-                size_t elementCount = typeDesc->getSize();
-                elementCount--;
-                size_t count = 0;
-                for (size_t i = 0; i < stringValues.size(); i++)
-                {
-                    if (count == elementCount)
-                    { 
-                        valueString += stringValues[i];
-                        uiProperties.enumerationValues.push_back(Value::createValueFromStrings(valueString, elemType));
-                        valueString.clear();
-                        count = 0;
-                    }
-                    else
-                    {
-                        valueString += stringValues[i] + ",";
-                        count++;
-                    }
-                }
-            }
-            else
-            {
-                uiProperties.enumerationValues.push_back(Value::createValue(enumerationValues));
-            }
-            propertyCount++;
-        }
-    }
-
-    const string& uiMinString = nodeDefElement->getAttribute(ValueElement::UI_MIN_ATTRIBUTE);
-    if (!uiMinString.empty())
-    {
-        ValuePtr value = Value::createValueFromStrings(uiMinString, nodeDefElement->getType());
-        if (value)
-        {
-            uiProperties.uiMin = value;
-            propertyCount++;
-        }
-    }
-
-    const string& uiMaxString = nodeDefElement->getAttribute(ValueElement::UI_MAX_ATTRIBUTE);
-    if (!uiMaxString.empty())
-    {
-        ValuePtr value = Value::createValueFromStrings(uiMaxString, nodeDefElement->getType());
-        if (value)
-        {
-            uiProperties.uiMax = value;
-            propertyCount++;
-        }
-    }
-    return propertyCount;
-}
-
-unsigned int getUIProperties(const string& path, DocumentPtr doc, const string& target, UIProperties& uiProperties)
-{
-    ValueElementPtr valueElement = findNodeDefChild(path, doc, target);
-    if (valueElement)
-    {
-        return getUIProperties(valueElement, uiProperties);
-    }
-    return 0;
-}
-
 } // namespace MaterialX
