+++ source/MaterialXRender/Util.h
@@ -12,6 +12,9 @@
 #include <MaterialXCore/Document.h>
 #include <MaterialXGenShader/ShaderGenerator.h>
 #include <MaterialXGenShader/GenContext.h>
+#include <MaterialXGenShader/Util.h>
+
+#include <map>
 
 namespace MaterialX
 {
@@ -28,7 +31,7 @@ namespace MaterialX
                                    const Color3& color);
 
     /// @}
-    /// @name Introspection utilities
+    /// @name Introspection and UI utilities
     /// @{ 
 
     /// Get a named uniform block for a given shader stage
@@ -36,6 +39,54 @@ namespace MaterialX
 
     /// Find a variable in shader's variable block which matches a given Element path.
     ShaderPort* findUniform(const VariableBlock* block, const std::string& path);
+
+    /// Set of possible UI properties for an element 
+    struct UIProperties
+    {
+        /// UI name
+        string uiName;
+
+        /// UI folder
+        string uiFolder;
+
+        /// Enumeration
+        StringVec enumeration;
+
+        /// Enumeration Values
+        vector<ValuePtr> enumerationValues;
+
+        /// UI minimum value
+        ValuePtr uiMin;
+
+        /// UI maximum value
+        ValuePtr uiMax;
+    };
+
+    /// Get the UI properties for a given nodedef element.
+    /// Returns the number of properties found.
+    unsigned int getUIProperties(const ValueElementPtr nodeDefElement, UIProperties& uiProperties);
+
+    /// Get the UI properties for a given element path. If the path is to a node, a target
+    /// identifier can be provided.
+    /// Returns the number of properties found.
+    unsigned int getUIProperties(const string& path, DocumentPtr doc, const string& target, UIProperties& uiProperties);
+
+    /// Interface for holding the UI properties associated shader port
+    struct UIPropertyItem
+    {
+        std::string label;
+        ShaderPort* variable = nullptr;
+        UIProperties ui;
+    };
+
+    /// A grouping of property items by name
+    using UIPropertyGroup = std::multimap<string, UIPropertyItem>;
+
+    /// Utility to group UI properties items based on ELement group name.
+    /// Returns a list of named and unnamed groups.
+    void createUIPropertyGroups(const VariableBlock& block, DocumentPtr contentDocument, TypedElementPtr materialElement,
+                                const string& pathSeparator, UIPropertyGroup& groups,
+                                UIPropertyGroup& unnamedGroups);
 
     /// @}
